requirements.txt
google-genai
python-dotenv
opencv-python-headless
numpy
ultralytics
fastapi
uvicorn
python-multipart
requests
supabase
